feat: implement ProtectedActionUseCase and related protocols for executing protected security actions refactor: update SwiftLint configuration and improve test cases for feature flags and logging chore: add Xcode build and test tasks to Taskfile fix: update SPDX document namespace and version
