+++ .swiftlint.yml
@@ -69,10 +69,10 @@ file_header:
 #   warning: 60
 #   error: 100
 
-# Type body length
-type_body_length:
-  warning: 310 # Increased slightly to accommodate test files
-  error: 400
+# Type body length - commented out since disabled
+# type_body_length:
+#   warning: 310 # Increased slightly to accommodate test files
+#   error: 400
 
 # File length - commented out since disabled
 # file_length:
