+++ Taskfile.yml
@@ -682,6 +682,8 @@ tasks:
       - echo "🔬 Running standard QA suite..."
       - task: qa:fix
       - task: swift:test
+      - task: xcode:build
+      - task: xcode:test
       - task: swift:lint
       - task: yaml:lint
       - task: yaml:codecov:validate
